refactor: Clean the whole codebase
This would prepare the repo to accept an updated code. I updated the
whole code in another local folder without cloning this one initially.
This way, there would be a lot conflict if i decide to just merge
without cleaning and updating.
During this whole period of updating the code, i wasn't pushing to git.
This was due to fact that the git i created for the update local folder broke. There
were too many errors, so i just left it and continued coding