feat, refactor: Add a view post page with comments
This commit still needs a couple of changes and final touches. The idea
behind storing and also fetching posts and comments as two different
entities needs to be redefined. I need to save both as posts in general
but c=with comments having a field that indicates it being a comment and
then i run the whole fetching this same way.

They were a couple of code refactoring i this commit. Jumbled up
already. Would do better henceforth.

Also, i discovered the need to cache the scroll position when
routing to another page. I started working on this but there are still
issues that needs attention. I initially focused on just home page plus
viewPost page. Then, i discovered that this caching would be needed
virtually on all pages. So route dependent caching is the way forward.