feat, refactor: Implement page scroll caching
Just like i've stated in the last commit, i intend to implement a scroll
cache. I had different ideas in accomplishing this which and i tried a
couple of them. I discovered loopholes in them except this particular
process i implemented here. By leveraging useLocation's key and pathname
properties to store scroll distance whenever a navigation to another
page is about to take place and then on cominng back to the page, a
check is done to compare the key/pathname in order to restore the scroll
distance.
I've managed to contain this whole process into two components. The
NavigateWithScrollCcahe which listens for click on the actual navigating
link then runs the caching. The other is the ScrollCache component. This
is responsible for restoring the scroll distance based on the cached
data.
There are conditions in choosing either the key or pathname property in
caching the scroll. This is based on the page. so, this should be kept
in mind when making a new page use this feature. The condition might
need to be updated.
The scrollCache requires that the ref for the scroll element is passed.
For NavigatewithScrollCache, a useState hook is should be created in the
component that uses this NSWC. so whenever the link is clicked, the
state is updated. This state is passed to the NSWC as prop. Also a
function that performs the navigation needs to be passed as prop. Then,
a cleanUp function that sets the clicked state back to false is need.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>