refactor: Refine the scroll caching process
I attempted using the scroll cache components pair for sidebar routing,
then i discovered that the refs for the pages would not be accessible
because sidebar is not a child but sibling to these pages and since i'm
using useContext hook to get these page nodes, they would be undefined
inside the NavigateWithScrollCache component.
This led me to actually create a context at the top level i.e Layout
component, just so all refs can be passed down every other component.
I leveraged the useImperativeHandle hook to update the Layout page ref.
With this, accessing any page's ref inside the NWSC component is
straightforward so far you're within that page.

I then implemented the scroll caching for the navigation page. This made
me discover the number of processes invlolved in setting the scroll
cache for a particular page. They are;
- Inside the page's component, create a ref that represents the node of
  the scrolling element.
- usecontext hook to derive the "pageNodes" ref from the Layoutcontext.
- useImperativeHandle hook to attach this page's ref to the "pageNodes"
  ref.
- Wrap the whole page component with the ScrollCache compoennt and also
  pass the page's ref as prop to it.
- Inside the NWSC component, follow the pattern used in the useState,
  useEffect and useMemo hooks in adding this page's node.
- If this page requires to use location.key in caching the scroll, then
  you can update the condition in the handleClick useCallback hook
  function.

 I also implemented a page refresh by clicking a sidebar page route while
 still being the same page. I made a refresh state in the layout
 component and then it is passed down to the SidebarItem and ScrollCache
 components using the context api. I used the pageRefresh as a
 dependency in the useLayouteffect hook just so that it re-renders
 anytime i setRefreshPage to true in the SidebarItem component after
 each click that satisfies the condition. I used a cleanup function to
 return the refresh page state back to false.

Random refactoring also happened in this commit