feat: Implement page scroll to "active" post section
This feature is very much needed because it makes the UX of the app
better. When a user clicks to view a particular comment, the page
scrolls down to the clicked comment section such that this comment is at
the top, making it the first comment the user sees after routing.

This implementation came with a lot of issues due to the fact that i
have to derive the offsetTop of the "active" post and then
set the page's scrollTop to this value.

- First issue was trying to make the "active" post ref.current a
  dependency for useEffect. The ref itself does not change when i route
  to another comment because its more or less the same component while
  its current property has actually changed. So using just the ref as a
  dependency would not re-render the offsetTop latest value.
  My quick fix for this was to remove the dependency but i realized that
  this woud be too much rerendering and besides useEffect gives an
  ESlint error because i'm setting a state without any dependency.

- Using the postId parameter as the dependency also came with its own
  bugs. The ref was providing a stale value for ref.current.offsetTop.
  Apparently, it takes some time to setup a attach a new node to the
  ref.
  I discovered you could pick nodes by passing a function to the
  ref attribute of the element and then the first parameter of that
  function represents the node just like normal events. I made a
  callback function and passed it into the ref attribute. This fuction
  was setting a state with the offsetTop value. I still had previous
  issues with this after making the postId the callback dependency. So,
  i resorted to timeout function. i used a 200ms time. The time was just
  right after my manual test.
  I converted this process into a react hook.
I passed the offsetTop generated as prop to the scrollCache to handle
the scroll process.

Another fix in this commit was fetching both current comment and parents
even when they've been loaded and cached as comments on parent post
load. This was done to take care direct page load on browser instead of
the normal navigation within the app.

Minor refactoring also occured here.