refactor: Modify the notification page
I created a proper schema together with the notifications document in
the the JSON db and each of these notifications have a relationship to
the users(subscribers) who should get the notification.
I craeted rtk query endpoint to handle notification fetch with the skip
limit process.
Modified the notification and notification block components to make use
of this updated route in fetching the notifications.
Provided the homeUserAvatar with the profileImage src as required.

Other minor refactoring also went down.