refactor: Ensure notification map is time based
Previously, i used the notification field in the authUser document to
map the notificationBlock. I did this so I could pass the viewed
property easily as prop to the NotificationBlock component.

Now, here comes the issue of mapping based on time in descending order.
The notification field doesn't include time so, I refactored my code to
make the selectIds from the adaptor as the array to map and then within
the NotificationBlock, the viewed property is generated from the
authUser. Since, the adapter already sorts the notitification based on
time.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>