refactor: Modified user fetch in profile page
I updated the profile page and its children components  to follow the
new process of fetching users.

Also, in homeUserAvatar, previous src where cached and on routing to a
new page(profile), it retains the former image instead of updating it.
I dealt with this by using a useEffect hook to enforce the src change.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>