fix: Fix profile routing on avatar click in EUL
The issue was that clicking a user avatar does not route to their
profile page but just closes the EngagedUsersList component.

In the EngagedUsersList component, the NavigateWithScrollCache component
deals with it navigation to profile page after a user avatar is clicked.
So, the function handleClick performs two operations; set the route
state to true, and also close down the EUL. It turns out that the EUl
gets closed down immediately even before the NWSC components performs
the actual routing.

To fix this, i had to pass this close action into the NWSC. I then
called the function immediately after navigating.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>