feat: Initiate the process of adding a story feat
The database schema and fetching endpoints have been defined here. The UI
for the story on the home page has changed. A story page has been added
also where the actual story is viewed.
The story page requires the use of components that makers use of the
LayoutContext but the story page is not an outlet of this Layout
component. I had to move the context into the Router component which is
like the topmost component. This sorts the issue.
This commit also contains some other refactoring.

Have all of these jumbled up changes in one commit is not good practice
and i'm sorry about that. Improvements would be made in future commits.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>