refactor: Change way of determining a viewed user
Previously, i designed my database in such a way that the viewed property
is placed alongside the storyAuthor in the otherStoryAuthors field of
the uaer document.
With this commit, the viewed property has been removed and the viewed
state is determined by confirming that all stories for a particular user
has been viewed.