refactor: Create a top level component
The need for a top level component was imminent. This component ensures
that the authenticated user has been fetched from the backend before
loading up other components.
This components also handles all properties that would be required
throughout the whole app.
The layout component used to serve this purpose until i started creating
the story feature which wasn't an outlet(child) of this component which makes
it impossible to have access to its context. Then, i moved the context
details to the router component which is ideal for the purpose. Now, i
realized that the router component is now handling more than its actual
functionality which is to route. I needed another component strictly for
handling the general details.