refactor: Ensure single component handles overlay
The transparent overlay was rendered when any of the popup it requires
has a "true" state (i.e rendered). with this commit, the transparent
component renders the popups and they are positioned based on their
click position. This improves the general code because there's a single
siource of truth now.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>