feat: Update the story feature
The update to the story feature involves refactoring other codes. The
transparent overlay was totally reformed into the same structure as the
opaque overlay just so there's a single source of truth. The story
options has been implemented here.

The preview design has been brought to life and further improvements
would still be made. There's a new component and also a cyustom hook for
the zoom feature. I also updated the editProfileImage component with
this zoom component just so there's uniformity.

There are a couple of other refactoring in this commit.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>