fix: Improve userCameo to cater for all use cases
The content of the userCameo component was flowing out of the screen on
resizing to a smaller screen in the story and block outlets of the
settings route. The width style needed a slight adjustment and also i
realized i had passed the avatar size prop as a string. I corrected this
by converting the prop in the userCameo component to avoid any error.

Author - Ahmed Quamordeen Gbolahan
gbolahanahmed1@gmail.com