feat: Improve the routes and controllers
- Create the stories route
- Create a getStories controller and attach it to the route.

- Create a notification route
- Create addNotification(for testing) and getNotifications controller
  and attach them to respective controllers.

While adding a new post, it could be regular post or a comment. The
regular post was the only one i catered for in previous commit. For a
comment type, the parents need to be passed along with the body. The
last item in the parents array would be the post that's being commented
on. This post's comment would be updated with this comment.

Notifications needs to be sent to concerned users after every post
regardless of whether its a regular post or a comment.
The process is as commented just below the addNewPost route