feat: Allow post reactions.
Liking and retweeting a post requires a mutation in the rtk query
endpoint. This reactions needs to updated immediately on the interface
without having to wait for a response from the backend.

I achieved this using the optimistic update feature of the rtk query.
This can be found within the reactToPost mutation endpoint. The
updateQueryData requires the query endPoint name for each post and also
its arguments in order to make adequate updates.

Infinite scroll is implemented which means that the fetch arguments
changes frequently since its handled in a state. These arguments cannot
be passed directly to the updateQueryData. For this reason, I attached
the both the endpoint name and arguments to the result from every query.
This way, the details are maintained and can be easily accessed.