feat: Implement the mutation of adding a new post
The invalidatesTag does not get the jpb done because of the fact that it
totally erases previous request on that endpoint. Also, there is a
flicker on the page whenever a request is made.

So, the refetch key function in the mutation result handled the
situation. Even when a cache of the request is still available, there is
a new call.

A couple modifications was made on some components to get this
running