feat: Implement create-post and delete-post
InvalidatesTags just wouldn't work seamlessly. There is always a few
seconds where the screen is blank due to post refetch. We do not want
this. So, the refetch method from the query result is being used here.
After sending the addPost request, on success, the skip/limit is updated
and then refetch is called. This makes the post update better.

Since the data for my post list is a merge of results from various
requests on the particular endpoint which are separated by skip/limit
values, some data becomes redundant. This means when you delete a post
and refresh the post lists, the post still remains on your page because
it still remains a part of the result data from earlier request(s).
To fix this, I leveraged the removePost redux state/action to blacklist
required post Ids. This way, the post never shows up on the page after
deletion. Ensured it was after a successful delete request that the
blacklisting process took place.

Other Minor modifications were made in this commit