feat: Display skeletons while notification loads
While fetching notifications, display skeletons for each notification in
anticipation.
After notifications are fetched, the userId is gotten from the result
and then this userId is used to fetch the whole user document. While
fetching this user, maintain skeleton on screen. If you do not do this,
there would be a short period of blank screen while it loads.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>