Fix: Implement quick response to post reactions.
Until this commit, there was a lag period whenever a post that has been
fetched on more than one endpoint is reacted to. i.e like or repost.

This commit ensures that the state contains only one cached record per
postId.

- Utilize the getTransform method to modify the response whenever a postId
has already been cached. The id of such post is changed into a random
and a cachedId property is attached to this post which represents the
actual postId that have been cached.
- In the post excerpt, The cachedIds are used to render whenver it is
truthy. Else, the normal id is used. The repostId remains the initial
option.

Optimistic update runs faster this way as it is only mutating a single
endpoint.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>