fix: Ensure the hiddenPost margin is balanced
The margin bottom for the hidden post seemed to be larger than the set
value. Turns out that the "post-container: last-of-type" had a margin
bottom value of 1. For some reasons, the hidden post is seen as the last
type for that classname regardless of its position on the stack. I
removed this style.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>