fix: Ensure that users can write on a new line
With the content-editable div, whenever you go to a new line and try to
write, the cursor automatically goes back to first line. This problem
was fixed by totally removing the div and now using the text area with a
JS code to ensure that the number of rows increases with the
introduction of a new line.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>