fix: Correct text area problem
problem: From the previous code, whenever new lines are added to the text area
either as word wrap or an outright new line by the Enter key, the text
area expandsn to fit its content. The case is different a case where a
new line is removed. The text area does not collapse but remain at same
height. This is not the behaviour that is expected.

solution: Returned to using the content editable div.
Controlling this div the react way by reassigning the innerHTML value
from the onInput event to the div's node caused a lot of bugs, so it was
left uncontrolled. This works fine by the way.

A content editable div wraps every new line in a new div tag. Therefore,
I replaced every opening div tag by a new line "\n" and the closing tag
by an empty string. This way, i'm able to format the innerHTML to
standard text.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>