fix: Improve the design of postExcerpt
The font size of the post text is increased.
The vsibleFor is removed from the fullDate
The fontSize of the full date is reduced to ensure a better look.
The createPost active top value was reduced just so that the it remains
within the viewport when a media with large height is selected.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>