fix: Improve UI of comment page
The offset of the "active comment" was 68px just so a part of the parent
post is seen to show the user that there are parents. I realized this
was not necessary as the scroll bar does the job of informing the user.
The offset is reduced to 60px.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>