fix: Avoid lag due to non-serializable media data
When a media file of a large size is uploaded, its base64 URl encoding
is  stored in a redux state. At this point, no error. Then, when the
createPost request is sent to the backend, the error comes up and the
app lags or even crashes depending on the file size.

Now, I ensured to remove this media file immediately in the function
that sends the request to the backend. This completely stops the lag or
crash. But, the serialization error still come up whenever a large file
size is uploaded and stored in the redux state. NO lag though, so i wave
off the error for now. I might just have to reduce the  file size limit
to one that wouldn't show the error.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>