feat: Implement infinite scroll properly
The batch fetch of content was done by using a timeout to request for
the next batch. With this commit, next batch of content is fetched
whenever a user reaches the end of a page. The proper way of
implementing infinite scroll.
The scrollCache component handles this. The fetchMore function is passed
to it to be triggered whenever the conditions have been met.
The timeout in the scroll event listener is to take care of the issue of
the fetchMore action being triggered more than once immediately the
condition is satisfied

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>