refact: Make report independent of post or story
Since its the same report component being used for story and posts, this
commit ensures that the reporting structure becomes independent and the
report component can be used anywhere by calling it using a single set
of action and also the state properties are gotten from a single source
which is in the layout slice.

Reporting progress would be tracked and for the confirmation pop-up to
be displayed. Hence, the isReporting and isReported properties in the
report state.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>