fix: Create separate endpoint 4 images and details
issue: The whole page reloads  after a successful request when the body
includes regular profile details and not image. No page reload whenever
its just avatar or cover photo that is to be updated.

solution: Have separate mutation endpoint for images and details. Use
optimistic update for the details. This would be done in next commit