feat: Initiate view story request on story exit
The useEffect return function comes to the rescue here.
This function gets triggered the first time the page is loaded. This is
not a desirable reaction. Therefore, a boolean state is created to
ensure that the function is not called on first render.

Author - Ahmed Quamordeen Gbolahan <gbolahanahmed1@gmail.com>